requirements.txt
smolagents[gradio,telemetry,litellm,mcp]==1.24.0
mcp
langfuse>=2.0,<3.0
openinference-instrumentation-smolagents
opentelemetry-sdk
opentelemetry-exporter-otlp
python-dotenv
pytest
